Utility functions for single blocks

These internal functions are the backend of Z2Block, Z2RowVecBlock and Z2ColVecBlock. They are unsafe, which means that they assume the blocks are of full size and do not check the validity of the arguments.

Z2Algebra.bitstringblock — Function

bitstringblock(x::Union{UInt64, Int64})

Convert a 64-bit integer to a block of bits in a human-readable format.

Example

julia> bitstringblock(TRIU_MASK) |> print
11111111
01111111
00111111
00011111
00001111
00000111
00000011
00000001

Z2Algebra.blockgetindex_mask — Function

blockgetindex_mask(i, j)

Get the mask where the bit(s) at 0-based position (i, j) in are 1. This function is low-level, and the caller is responsible for bounds checking.

Note

Unlike blockgetindex, this function does not shift bits to the least significant position.

Z2Algebra.blockgetindex — Function

blockgetindex(x::UInt64, i, j)

Get the bit(s) at 0-based position (i, j) in the block. This function is low-level, and the caller is responsible for bounds checking.

Note

Unlike blockgetindex_mask, this function shifts bits to the least significant position.

Z2Algebra.rowvecones — Function

rowvecones(len::Integer)::UInt64

Backend for ones(Z2Number, len) that yields a row vector. See also colvecones and blockones.

Z2Algebra.colvecones — Function

colvecones(len::Integer)::UInt64

Backend for ones(Z2Number, len) that yields a column vector. See also rowvecones and blockones.

Z2Algebra.blockones — Function

blockones(m::Integer, n::Integer)::UInt64

Backend for ones(Z2Number, m, n) that yields a block. See also rowvecones and colvecones.

Z2Algebra.column2row — Function

column2row(x::UInt64)::UInt64

Transposes the block where only the first column is nonzero.

Z2Algebra.row2column — Function

row2column(x::UInt64)::UInt64

Transposes the block where only the first row is nonzero.

Z2Algebra.transposeblock — Function

transposeblock(x::UInt64)::UInt64

Transposes the block.

Z2Algebra.getcolsum — Function

getcolsum(x::UInt64)::UInt64

Calculates the sum of each column and returns it in the first row of the returned block.

Z2Algebra.getrowsum — Function

getrowsum(x::UInt64)::UInt64

Calculates the sum of each row and returns it in the first column of the returned block.

Z2Algebra.getallsum — Function

getallsum(x::UInt64)::UInt64

Calculates the sum of all bits in the block and returns it in the top-left bit of the returned block.

Z2Algebra.colvecmulrowvec — Function

colvecmulrowvec(x::UInt64, y::UInt64)::UInt64

Calculates the product of a column vector and a row vector.

Z2Algebra.matmulmat — Function

matmulmat(x::UInt64, y::UInt64)::UInt64

Multiplies two blocks.

Z2Algebra.matmulrowvec — Function

matmulrowvec(x::UInt64, y::UInt64)::UInt64

Multiplies a block by a row vector.

Z2Algebra.matmulcolvec — Function

matmulcolvec(x::UInt64, y::UInt64)::UInt64

Multiplies a block by a column vector.

Z2Algebra.matldivmat — Function

matldivmat(x::UInt64, y::UInt64)::UInt64

Left divides two blocks.

Z2Algebra.padidentity — Function

padidentity(x::UInt64, p)::UInt64

Assuming x is a square block, pad an identity matrix to the bottomright to form an 8x8 block. p is the size of x.

Z2Algebra._to_vec_block — Function

_to_vec_block(dest<:Z2Vector, src<:Z2Vector, v::Z2Block)

Convert a Z2Block v of a src type to a block of dest type.