Internals

General

Z2Algebra.check_z2array_valid — Function

check_z2array_valid(::Union{Z2RowVector,Z2ColVector,Z2Matrix})

Check if the internal data is consistent. The check should pass unless you use the internal constructors.

Z2Algebra.tailmask! — Function

tailmask!(::T) -> T
tailmask!(T, blocks, tailsize)

Make a Z2Array valid by masking the blocks with tailsize information.

LUutils

Z2Algebra.LUutils.blockswaprows — Method

blockswaprows(a::Z2Block, maski, maskj, shift)

Swap two rows in a block and return the result block. maski and maskj are row masks generated by blockgetindex_mask, and shift is the bit difference between the masks. The arguments are set in this way because maski, maskj and shift could be reused in swaprows(@ref).

Z2Algebra.LUutils.blockswaprows — Method

blockswaprows(a::Z2Block, b::Z2Block, maski, maskj, shift)

Swap two rows in two blocks and return the result blocks.

Z2Algebra.LUutils.eliminate_rows! — Method

eliminate_rows!(blocks::AbstractMatrix{Z2Block}, blockpivot, pivot, colbuffer::AbstractVector{UInt64})

For the pivot with index (blockpivot, pivot), do a round of row elimination. colbuffer is pre-allocated to reduce allocation overhead, and its length should be at least size(blocks,1).

Z2Algebra.LUutils.find_first_nonzero — Method

find_first_nonzero(blocks, blockcol, col)

Find the first non-zero element in the column col of the block matrix blocks starting from block blockcol. Returns (blockrow, row). Returns (0, 0) if the element is not found.

Z2Algebra.LUutils.swaprows! — Method

swaprows!(blocks, blockcol, col, blockrow, row)

Swap two rows in blocks in the LU process. blockcol, col, blockrow and row are the indices of the pivot. The pivot and the entries to its right should be swapped to the row such that the pivot is in the diagonal.